Fixes for nodes / edges, solve circuit (Details) - Inductor no longer has polarity - Fix for directly adjacent nodes in branch finding - Branch fix for nodes close to each other - Ignore resistance in voltage branches - Fix some abs() errors in gcc-6 (Unrelated to circuits) - Fix visited not updating sometimes causing node ids to be offset - Add polarity PSCN / NSCN requirement for CAPR and VOLT

+++ src/simulation/Air.cpp
@@ -76,7 +76,7 @@ void Air::update_airh(void)
 		for (x=0; x<XRES/CELL; x++)
 		{
 			// Apply time dilation: (negative = loop every n frames)
-			if (sim.time_dilation[y][x] < 0 && sim.timer % abs(sim.time_dilation[y][x]) != 0)
+			if (sim.time_dilation[y][x] < 0 && sim.timer % (int)abs(sim.time_dilation[y][x]) != 0)
 				continue;
 			int count = 0;
 			air_cell_update_begin:
@@ -225,7 +225,7 @@ void Air::update_air(void)
 			for (x=0; x<XRES/CELL; x++)
 			{
 				// Apply time dilation: (negative = loop every n frames)
-				if (sim.time_dilation[y][x] < 0 && sim.timer % abs(sim.time_dilation[y][x]) != 0)
+				if (sim.time_dilation[y][x] < 0 && sim.timer % (int)abs(sim.time_dilation[y][x]) != 0)
 					continue;
 				int count = 0;
 				air_cell_update_begin:
